requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# color-to-swatch.py: 2
Click == 7.0

# color-to-swatch.py: 5
colour == 0.1.5

# color-to-swatch.py: 9
python_slugify == 1.2.6

# color-to-swatch.py: 3,4
selenium == 3.14.1
